+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_correct_test_file_name/config_parser.dart
@@ -1,0 +1,11 @@
+part of 'prefer_correct_test_file_name_rule.dart';
+
+class _ConfigParser {
+  static const _namePatternConfig = 'name-pattern';
+
+  static String parseNamePattern(Map<String, Object> config) {
+    final raw = config[_namePatternConfig];
+
+    return raw is String ? raw : '_test.dart';
+  }
+}