+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_correct_test_file_name/visitor.dart
@@ -1,0 +1,24 @@
+part of 'prefer_correct_test_file_name_rule.dart';
+
+class _Visitor extends GeneralizingAstVisitor<void> {
+  final String path;
+  final String pattern;
+
+  final _declarations = <FunctionDeclaration>[];
+
+  Iterable<FunctionDeclaration> get declaration => _declarations;
+
+  _Visitor(this.path, this.pattern);
+
+  @override
+  void visitFunctionDeclaration(FunctionDeclaration node) {
+    // ignore: deprecated_member_use
+    if (node.name.name != 'main' || _matchesTestName(path)) {
+      return;
+    }
+
+    _declarations.add(node);
+  }
+
+  bool _matchesTestName(String path) => path.endsWith(pattern);
+}