+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_redundant_async/examples/example.dart
@@ -1,0 +1,33 @@
+Future<void> usesAwait(Future<String> later) async {
+  print(await later);
+}
+
+Future<void> asyncError() async {
+  throw 'Error!';
+}
+
+Future<void> asyncValue() async => 'value';
+
+// LINT
+Future<int> fastestBranch(Future<int> left, Future<int> right) async {
+  return Future.any([left, right]);
+}
+
+class SomeClass {
+  void syncMethod() {}
+
+  Future<int> asyncMethod() async => 1;
+
+  // LINT
+  Future<String> anotherAsyncMethod() async => Future.value('value');
+
+  Future<String> someAsyncMethod(Future<String> later) => later;
+
+  Future<void> report(Iterable<String> records) async {
+    if (records.isEmpty) {
+      return;
+    }
+
+    print(records);
+  }
+}